deleted some unnecessary things for cleanup
