+++ frontend/src/components/Posts.jsx
@@ -3,41 +3,26 @@ import { usePostContext } from "../context/PostContext"
 import SinglePost from "./SinglePost"
 
 function Posts() {
-    const comics = [
-        {
-            id: 1,
-            title: "Test 1",
-            image: "https://placehold.co/600x400?text=Comic+Strip+1",
-            alt: "placeholder comic",
-            caption: "Some example text to fill out this section"
-        },
-        {
-            id: 2,
-            title: "Test 2",
-            image: "https://placehold.co/600x400?text=Comic+Strip+2",
-            alt: "placeholder comic 2",
-            caption: "This is where the user posts their caption"
-        }
-    ]
     const [errorMessage, setErrorMessage] = useState("");
-
     const {currentPosts, handleUpdatePosts} = usePostContext()
 
-    const posts = async() => {
+    const loadPosts = async() => {
         try {
             const response = await fetch("http://localhost:8000/api/posts")
                 .then((response) => response.json())
-            handleUpdatePosts(response.data)
+            const newPosts = response.data
+            handleUpdatePosts(newPosts)
         } catch (error) {
             setErrorMessage("Something seems to be wrong. Try again");
             console.error("Oops", error);
         }
     }
-    const reload = setInterval(posts, 10_000)
+    const reload = setInterval(loadPosts, 10_000)
     reload
+    
     const postList = currentPosts.map(post => (
         <SinglePost
-            key={post.id}
+            key={post._id}
             title={post.title}
             image={post.image}
             alt={post.alt}
