+++ frontend/src/forms/NewPostForm.jsx
@@ -1,52 +1,108 @@
 import { useState } from "react";
+import { Link, useNavigate } from "react-router-dom";
 import { usePostContext } from "../context/PostContext";
 
-export default function NewPostForm({onAddPost}) {
-    const [title, setTitle] = useState('')
-    const [image, setImage] = useState('')
-    const [alt, setAlt] = useState('')
-    const [caption, setCaption] = useState('')
-    const [submitResult, setSubmitResult] = useState('')
+function NewPostForm() {
+  const [title, setTitle] = useState("");
+  const [image, setImage] = useState("");
+  const [alt, setAlt] = useState("");
+  const [caption, setCaption] = useState("");
 
-    const {currentPosts, handleUpdatePosts} = usePostContext()
+  const navigate = useNavigate();
 
-    const handleSubmit = (e) => {
-        e.preventDefault()
-        const data = new FormData(e.target)
+  const [response, setResponseMessage] = useState("");
+  const [errorMessage, setErrorMessage] = useState("");
+  const {currentPosts, handleUpdatePosts} = usePostContext()
 
-        const newPost = Object.fromEntries(data)
-        console.log(newPost)
-        onAddPost({newPost})
-        handleUpdatePosts([...currentPosts, newPost])
-        setSubmitResult('Successful post')
+  const handleSubmit = async (e) => {
+    e.preventDefault();
+
+    const postData = {
+      title,
+      image,
+      alt,
+      caption,
+    };
+
+    try {
+      const response = await fetch("http://localhost:8000/api/posts/create", {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify(postData),
+      });
+
+      const postList = await fetch("http://localhost:8000/api/posts")
+        .then((response) => response.json())
+        // .then((json) => {
+        //     console.log(json.data)
+        // });
+      
+      const data = await response.json();
+      if (response.ok) {
+        console.log(data.data)
+        console.log(postList.data)
+        handleUpdatePosts(postList.data)
+        setResponseMessage("Success");
+        navigate("/home");
+      } else {
+        setErrorMessage(data.error);
+        console.error("Oops: ", data.error);
+      }
+    } catch (error) {
+      setErrorMessage("Something seems to be wrong. Try again");
+      console.error("Oops", error);
     }
+  };
 
-    return (
+  return (
+    <div>
+      <form onSubmit={handleSubmit}>
         <div>
-            <form onSubmit={handleSubmit}>
-                <div>
-                    <label>Title: 
-                        <input name="title" value={title} onChange={e => setTitle(e.target.value)} />
-                    </label>
-                </div>
-                <div>
-                    <label>Image: 
-                        <input name="image" value={image} onChange={e => setImage(e.target.value)} />
-                    </label>
-                </div>
-                <div>
-                    <label>Alternative Text: 
-                    <input name="alt" value={alt} onChange={e => setAlt(e.target.value)} />
-                </label>
-                </div>
-                <div>
-                    <label>Caption: 
-                    <input name="caption" value={caption} onChange={e => setCaption(e.target.value)} />
-                </label>
-                </div>
-                <button>Post</button>
-            </form>
-            <div>{submitResult}</div>
+          <label>
+            Title:
+            <input
+              name="title"
+              value={title}
+              onChange={(e) => setTitle(e.target.value)}
+            />
+          </label>
         </div>
-    )
+        <div>
+          <label>
+            Image:
+            <input
+              name="image"
+              value={image}
+              onChange={(e) => setImage(e.target.value)}
+            />
+          </label>
+        </div>
+        <div>
+          <label>
+            Alternative Text:
+            <input
+              name="alt"
+              value={alt}
+              onChange={(e) => setAlt(e.target.value)}
+            />
+          </label>
+        </div>
+        <div>
+          <label>
+            Caption:
+            <input
+              name="caption"
+              value={caption}
+              onChange={(e) => setCaption(e.target.value)}
+            />
+          </label>
+        </div>
+        <button>Post</button>
+      </form>
+    </div>
+  );
 }
+
+export default NewPostForm;
